This the main repository for the IndividuAI project
The project dives in the development of an "AI Personalized Chat bot" - IndividuAI, a cutting-edge Software as a Service (SaaS) platform. This platform, integrating artificial intelligence chatbot technology and financial technology (FinTech), focuses on delivering tailored interactions based on user input and data, with a special emphasis on subscription payment methods for software services.